chore: refactor duplicate logic
