+++ packages/shared/lib/contexts/onboarding/actions/index.ts
@@ -11,6 +11,7 @@ export * from './initialiseOnboardingProfile'
 export * from './initialisePincodeManager'
 export * from './initialiseProfileManagerFromOnboardingProfile'
 export * from './initialiseSecretManager'
+export * from './resetOnboardingProfile'
 export * from './resetOnboardingProfileWithProfileManager'
 export * from './resetOnboardingProfileWithAlreadyStoredMnemonic'
 export * from './restoreBackupFromStrongholdFile'
