+++ packages/shared/lib/contexts/onboarding/actions/resetOnboardingProfile.ts
@@ -1,0 +1,15 @@
+import {
+    deleteOnboardingProfile,
+    initialiseOnboardingProfile,
+    onboardingProfile,
+    updateOnboardingProfile,
+} from '@contexts/onboarding'
+import { get } from 'svelte/store'
+
+export async function resetOnboardingProfile(): Promise<void> {
+    const onboardingProfileData = get(onboardingProfile)
+    await deleteOnboardingProfile()
+    initialiseOnboardingProfile(onboardingProfileData?.isDeveloperProfile)
+    const { id } = get(onboardingProfile)
+    updateOnboardingProfile({ id, ...onboardingProfileData })
+}
