+++ packages/shared/lib/contexts/onboarding/actions/resetOnboardingProfileWithProfileManager.ts
@@ -1,23 +1,15 @@
-import {
-    deleteOnboardingProfile,
-    initialiseOnboardingProfile,
-    IOnboardingProfile,
-    onboardingProfile,
-    updateOnboardingProfile,
-} from '@contexts/onboarding'
 import { get } from 'svelte/store'
+
+import { onboardingProfile, updateOnboardingProfile } from '../stores'
+
+import { resetOnboardingProfile } from './resetOnboardingProfile'
 
 export async function resetOnboardingProfileWithProfileManager(): Promise<void> {
     if (!get(onboardingProfile)?.hasInitialisedProfileManager) {
         return
     } else {
-        const onboardingProfileData: Partial<IOnboardingProfile> = { ...get(onboardingProfile) }
-        await deleteOnboardingProfile()
-        await initialiseOnboardingProfile(onboardingProfileData?.isDeveloperProfile)
-        const { id } = get(onboardingProfile)
+        await resetOnboardingProfile()
         updateOnboardingProfile({
-            ...onboardingProfileData,
-            id,
             type: null,
             recoveryType: null,
             hasInitialisedProfileManager: false,
